fixed data type of variable i in for loop.
forgot that the .size() function returns a value of type size_t, so needed to change from int -> size_t

+++ image_processor.cpp
@@ -63,7 +63,7 @@ std::vector<uint16_t> ImageProcessor::calculateDiff(const std::vector<uint16_t>&
 	std::vector<uint16_t> output;
 
 	// assuming both vectors are the same size.
-	for (int i = 0; i < imageA.size(); ++i) {
+	for (size_t i = 0; i < imageA.size(); ++i) {
 		//uint16_t difference = imageA[i] - imageB[i]; used this variable to test underflow
 		uint16_t result = std::abs(imageA[i] - imageB[i]);
 		output.push_back(result);
